Split signers from providers
