+++ src.ts/signers/index.ts
@@ -1,0 +1,3 @@
+export type { Signer } from './signer.js';
+
+export { AbstractSigner, VoidSigner } from './abstract-signer.js';
