+++ src.ts/signers/signer.ts
@@ -2,8 +2,8 @@ import type { Addressable } from '../address/index.js';
 import type { TypedDataDomain, TypedDataField } from '../hash/index.js';
 import type { TransactionLike } from '../transaction/index.js';
 
-import type { ContractRunner } from './contracts.js';
-import type { BlockTag, Provider, TransactionRequest, TransactionResponse } from './provider.js';
+import type { ContractRunner } from '../providers/contracts.js';
+import type { BlockTag, Provider, TransactionRequest, TransactionResponse } from '../providers/provider.js';
 
 /**
  * A Signer represents an account on the Ethereum Blockchain, and is most often backed by a private key represented by a
